fix: refactored handler into seperate class
